feat: big integer id field

+++ web/app/catalogs/models.py
@@ -136,7 +136,7 @@ class Settings(models.Model):
     name = models.CharField(" - ", max_length=50, default="Настройки", blank=True)
 
 
-    admin_id = models.IntegerField("Telegram Id (менеджер)", null=True, blank=True, default=None)
+    admin_id = models.BigIntegerField("Telegram Id (менеджер)", null=True, blank=True, default=None)
     main_channel_url = models.CharField("ссылка основного телеграм канала", default=" ", max_length=255, blank=True)
     help_url = models.CharField("ссылка поддержки", default=" ", max_length=255, blank=True)
     bot_url = models.CharField("ссылка бота", default=" ", max_length=255, blank=True)
